# unicorn-api/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    properties:
      details:
        description: |-
          Additional error details
          example: Field 'email' is required
        type: string
      error:
        description: |-
          The error message
          example: Invalid request
        type: string
      status_code:
        description: |-
          The HTTP status code
          example: 400
        type: integer
      timestamp:
        description: |-
          The timestamp when the error occurred
          example: 2024-01-01T12:00:00Z
        type: string
    type: object
...
  models.LambdaExecuteRequest:
    properties:
      process:
        $ref: '#/definitions/models.LambdaProcessInfo'
      project:
        properties:
          entry:
            example: |-
              import utils
              print(utils.add(1,2))
            type: string
          files:
            items:
              $ref: '#/definitions/models.LambdaFile'
            type: array
        required:
        - files
        type: object
      runtime:
        properties:
          name:
            example: python3
            type: string
          version:
            example: "3.12"
            type: string
        required:
        - name
        type: object
    type: object
  models.LambdaExecuteResponse:
    properties:
      output:
        example: |
          3
        type: string
      runtime:
        example: python3
        type: string
      status:
        example: success
        type: string
      time:
        example: 0.023s
        type: string
    type: object
  models.LambdaFile:
    properties:
      contents:
        example: "def add(a, b):\n\treturn a + b"
        type: string
      name:
        example: utils.py
        type: string
    required:
    - contents
    - name
    type: object
  models.LambdaPermissions:
    properties:
      network:
        example: false
        type: boolean
      read:
        example: true
        type: boolean
      write:
        example: false
        type: boolean
    type: object
  models.LambdaProcessInfo:
    properties:
      env:
        additionalProperties:
          type: string
        example:
          '{"DEBUG"': '"true"}'
        type: object
      max_opened_files:
        example: 10
        type: integer
      max_processes:
        example: 5
        type: integer
      permissions:
        $ref: '#/definitions/models.LambdaPermissions'
      stdin:
        example: test input
        type: string
      time:
        example: 2s
        type: string
    type: object
...
  models.SecretBodyRequest:
    properties:
      metadata:
        type: string
      name:
        type: string
      value:
        type: string
    required:
    - name
    - value
    type: object
  models.SecretResponse:
    description: A secret response without sensitive data.
    properties:
      created_at:
        type: string
      id:
        type: string
      metadata:
        type: string
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  models.UpdateSecretBody:
    properties:
      metadata:
        type: string
      value:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@unicorn-api.com
    name: Unicorn API Support
    url: https://github.com/your-org/unicorn-api
  description: A comprehensive RESTful API for Unicorn services providing Identity
    and Access Management (IAM) functionality. This API supports user authentication,
    role-based access control, organization management, and JWT token handling.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Unicorn API
  version: "1.0"
paths:
...
  /api/v1/lambda/execute:
    post:
      consumes:
      - application/json
      description: Execute code in a Lambda function with specified runtime and files
      parameters:
      - description: Lambda execution request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LambdaExecuteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LambdaExecuteResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Execute a Lambda function
      tags:
      - Lambda
  /api/v1/lambda/test:
    post:
      consumes:
      - application/json
      description: Test a Lambda function with specified runtime and files
      parameters:
      - description: Lambda test request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LambdaExecuteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LambdaExecuteResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Test a Lambda function
      tags:
      - Lambda
...
  /api/v1/secrets:
    get:
      description: List all secrets for the authenticated user (does not return secret
        values)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SecretResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List secrets
      tags:
      - Secrets Manager
    post:
      consumes:
      - application/json
      description: Create a new secret for the authenticated user
      parameters:
      - description: Secret details to create
        in: body
        name: secret
        required: true
        schema:
          $ref: '#/definitions/models.SecretBodyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SecretResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create secret
      tags:
      - Secrets Manager
  /api/v1/secrets/{id}:
    delete:
      description: Delete a secret for the authenticated user
      parameters:
      - description: Secret ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete secret
      tags:
      - Secrets Manager
    get:
      description: Get a secret and its decrypted value for the authenticated user
      parameters:
      - description: Secret ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get secret
      tags:
      - Secrets Manager
    put:
      consumes:
      - application/json
      description: Update the value and/or metadata for a secret
      parameters:
      - description: Secret ID
        in: path
        name: id
        required: true
        type: string
      - description: Secret update
        in: body
        name: secret
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSecretBody'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update secret
      tags:
      - Secrets Manager
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"